DBZにも存在するクラスについて削除。 DBCビジネスのテストを一部修正。

+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/ShokanJuryoininKeiyakusha.java
@@ -67,7 +67,7 @@ public class ShokanJuryoininKeiyakusha extends ModelBase<ShokanJuryoininKeiyakus
                 entity.getHihokenshaNo(),
                 entity.getShoKisaiHokenshaNo(),
                 entity.getUketsukeYMD(),
-                entity.getRirekiNo());
+                entity.getRirekiNo().intValue());
     }
 
     /**
